🌈feat : 유저 정보 수정 새로운 디자인으로 변경, 리팩토링
