+++ src/pages/Main/index.tsx
@@ -4,57 +4,64 @@ import MainBody from '@components/Main/MainBody/MainBody';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useSpaceInfoQuery } from '@hooks/api/space/useSpaceInfoQuery';
-import { useWelcomeModal } from '@/store/modal';
-import WelcomeAndSettingModal from '@/components/Main/WelcomeAndSettingModal';
-import ImgEditModal from '@/components/Create/ImageEditModal/ImageEditModal';
-import { ImageArrType } from '@/types/image.type';
-
+import WelcomeModal from '@/components/Main/WelcomeAndSettingModal/WelcomeModal';
+import UserSettingModal from '@/components/Main/WelcomeAndSettingModal/UserSettingModal';
+import { Modal } from '@mui/material';
 const MainPage = () => {
   const [selectState, setSelectState] = useState(false);
   const spaceId = useParams().spaceId!;
   const spaceInfo = useSpaceInfoQuery(spaceId).spaceInfo!;
   const { userList, tagList, isFirst, spaceTitle, imgUrl } = spaceInfo;
-  const [isImageModalOpen, setImageModalOpen] = useState(false);
 
-  const { isOpen: isWelcomeModalOpen, ModalOpen: WelcomeModalOpen } =
-    useWelcomeModal();
+  const [modalNum, setModalNum] = useState(0);
+
+  //세팅 모달로 이동
+  const onMoveSettingModal = () => {
+    setModalNum(2);
+  };
+
+  //모달 닫기
+  const onCloseSettingModal = () => {
+    setModalNum(0);
+  };
 
   //처음이라면 welcome 모달 열기
   useEffect(() => {
-    if (isFirst) {
-      WelcomeModalOpen();
+    console.log(isFirst); //이걸로 바꾸기
+    const fakeIsFirst = false;
+    if (fakeIsFirst) {
+      setModalNum(1);
     }
-  }, [isFirst, WelcomeModalOpen]);
-
-  //이미지 저장하는 곳
-  const [imageArr, setImageArr] = useState<ImageArrType>({
-    images: [],
-    cropImages: [],
-    convertedImages: [],
-  });
+  }, []);
 
   return (
     <>
-      {/*처음 스페이스 방문시 보이는 모달*/}
-      {isWelcomeModalOpen && (
-        <WelcomeAndSettingModal
-          spaceTitle={spaceTitle}
-          imgUrl={imgUrl}
-          setImageArr={setImageArr}
-          imageArr={imageArr}
-          setImageModalOpen={setImageModalOpen}
-          isImageModalOpen={isImageModalOpen}
-        />
+      {modalNum && userList && (
+        <Modal
+          open={modalNum > 0}
+          slotProps={{
+            backdrop: {
+              sx: {
+                backgroundColor: 'rgba(0, 0, 0, 0.6)',
+              },
+            },
+          }}
+        >
+          {modalNum === 1 ? (
+            <WelcomeModal
+              onMoveSettingModal={onMoveSettingModal}
+              spaceTitle={spaceTitle}
+              imgUrl={imgUrl}
+            />
+          ) : (
+            <UserSettingModal
+              userNames={userList.map((user) => user.userName)}
+              ModalClose={onCloseSettingModal}
+            />
+          )}
+        </Modal>
       )}
-      {/*이미지 편집 모달*/}
-      {isImageModalOpen && (
-        <ImgEditModal
-          imageArr={imageArr}
-          setImageArr={setImageArr}
-          isCircle={true}
-          setImageModalOpen={setImageModalOpen}
-        />
-      )}
+
       <S.Wrapper>
         <MainHeader
           spaceId={spaceId}
